refactor(all): rename CallModule to DataModule
